scripts/requirements.txt
gspread==5.12.0
oauth2client==4.1.3
PyYAML==6.0.1
